feat: implement custom hook to fetch authenticated user data
